extract `moment` class
