+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * refactor - explicit import and export types [#112](https://github.com/LoveCommunity/love.dart/issues/112)
 * refactor - remove unused internal type `CopySystem` [#113](https://github.com/LoveCommunity/love.dart/issues/113)
+* refactor - extract `moment` class [#114](https://github.com/LoveCommunity/love.dart/issues/114)
 
 ## [0.1.0-rc.2] - 2020-10-28
