+++ lib/src/types/types.dart
@@ -40,11 +40,3 @@ typedef Run<State, Event> = Disposer Function({
 });
 
 typedef CopyRun<State, Event> = Run<State, Event> Function(Run<State, Event> run);
-
-/// A holder of state, oldState and event of a moment
-class Moment<State, Event> {
-  final State state;
-  final State? oldState;
-  final Event? event;
-  const Moment(this.state, this.oldState, this.event);
-}
